sandbox compiles again and tests pass

+++ sandbox/src/main.rs
@@ -3,7 +3,7 @@ use std::iter::once;
 use std::sync::Arc;
 use std::time::Instant;
 
-use glam::{Quat, Vec3, vec3};
+use glam::{Quat, Vec3, dvec3};
 use image::DynamicImage;
 use image::codecs::hdr::HdrDecoder;
 use storm_controls::EguiControls;
@@ -43,7 +43,7 @@ fn create_shapes() -> ((Box3D, Transform), (Ball, Transform)) {
     let aab_transform = Transform::IDENTITY;
 
     let ball = Ball::from_radius(1.0);
-    let ball_transform = Transform::from_translation(vec3(2.0, 0.0, 0.0));
+    let ball_transform = Transform::from_translation(dvec3(2.0, 0.0, 0.0));
 
     ((aab, aab_transform), (ball, ball_transform))
 }
